requirements.txt
absl-py==1.0.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.5
cachetools==5.0.0
catboost==1.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cvlib==0.2.7
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.0.3
flatbuffers==2.0
fonttools==4.32.0
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.43.0
gTTS==2.3.0
h5py==3.6.0
idna==3.3
imageio==2.17.0
imbalanced-learn==0.9.1
importlib-metadata==4.12.0
imutils==0.5.4
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.9.0.1
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
nltk==3.8.1
notebook==6.4.11
numpy==1.22.2
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
playsound==1.3.0
plotly==5.10.0
progressbar==2.5
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0b3
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.2
scipy==1.8.0
seaborn==0.12.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.12.2
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.4.8
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.8
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wrapt==1.13.3
zipp==3.8.1
